+++ firmware/targets/f2/Src/flipper_hal.c
@@ -40,7 +40,7 @@ void delay_us_init_DWT(void) {
     DWT->CYCCNT = 0U;
 }
 
-void delay_us(uint32_t time) {
+void delay_us(float time) {
     uint32_t start = DWT->CYCCNT;
     uint32_t time_ticks = time * (SystemCoreClock / 1000000);
     while((DWT->CYCCNT - start) < time_ticks) {
